moves App to be one level about component
